+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This project uses [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- Added new test workflow for Github.
+
 ## 1.12.1 - 2026-08-11
 
 - Updated TOC for WoW 12.1.0.