+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project uses [Semantic Versioning](http://semver.org/).
 
 - Added new test workflow for Github.
 - Removed GuildRoster tests now that C_GuildInfo namespace is standard.
+- Added Lua version to version command output.
 
 ## 1.12.0 - 2026-08-02