+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project uses [Semantic Versioning](http://semver.org/).
 
 - Added new test workflow for Github.
 - Added Lua version to version command output.
+- Forward guild and officer messages sent directly by macros or addons.
 
 ## 1.12.1 - 2026-08-11